Add files via upload
Hi. This is my first time ever using GitHub. Currently I am working on a program that can solve the issue of construction steel cuts. The steel rods comes at a standard length of 12 meters, and that is cut according to the needed lengths in site. In most cases, the needed pieces are not consistent and could not be divided in a way that uses a complete rod. This results in short and useless pieces from every rod. These losses are wasted money. The cuts could be organized to minimize the losses. But even for the smallest amount of steel, it is very complex task. fortunately we have computers. 

So far I was able to create a code that could take a list of cuts, and isolated the cuts that forms a complete rod, 12 meters, and remove them from the original list. Next the code will take the new list and figure all the combinations that give less that 12 meters. I separated those into three groups; one with lists of all possible 4 cuts summed to less that 12 . Another with lists of 3 cuts and the last with pairs summed to less than 12. I stopped at the point of removing all the duplicates form the lists.